Workaround: Add Sort Priority to flow-directions
It looks like csv2rdf/ table2qb are mis-interpreting the columns
in the csv file (i.e. on the basis of position not title).

This change should make the data correct for now. Ultimately we'll
need to fix the pipeline to ensure that either 'Sort Priority' or
'Description' can be provided in the 4th column.